% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{rfold}
\alias{rfold}
\title{Transfer .R files into the R directory}
\usage{
rfold(folders_to_ignore = NULL, script_name_prefix = "DO_NOT_EDIT_")
}
\arguments{
\item{folders_to_ignore}{a string vector of plain names folders that should be ignored
where transferring .R files into the main R folder (for example a dev folder). Defaults to NULL}

\item{script_name_prefix}{characters string of length 1 that will be appended to the start of the .R scripts when transferring them to the R folder. Defaults to 'DO_NOT_EDIT', set it to NULL for no appending}
}
\value{
called for the side effect of transferring all R files
}
\description{
Transfer .R files into the R directory
}
